Split library code from executable
